// src/api/orval/model/telegramBotUserDto.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * OpenVPN Gate Monitor API
 * OpenAPI spec version: v1
 */

export interface TelegramBotUserDto {
  id?: number;
  telegramId?: number;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  isAdmin?: boolean;
  isBlocked?: boolean;
  createDate?: string;
  lastUpdate?: string;
}